Fix the error of zeros in denominator
This problem can occasionally occurs when there is at least one one
zero in the matrix after possible translation, scaling and zero-padding,
which happens when updating the permutation matrix in Eq. 28. The
original code adds the absolute value of minimum value, which is not
sufficient for matrix with zeros. Now, the matrix elements will add the
maximum of the absolute values of the matrices.