Refactor the file downloading section
The original block downloads the input files with the same directory
of the Jupyter files, but this is redundant if one clones the repo.
Moreover, if one runs the Procrustes analysis blocks without running
the download block, it will result to errors as the file path it points
to is not right. Now everything is fixed and running in Binder does not
require people to download the input files.